+++ src/services/myPage/postMyPageUserInfo.ts
@@ -1,6 +1,7 @@
 "use server";
 
 import {MyPageUserInfo} from "@/services/myPage/getMyPageUserInfo";
+import {cookies} from "next/headers";
 
 const apiUrl = process.env.NEXT_PUBLIC_API_URL;
 
@@ -18,21 +19,22 @@ export interface MyPageUserInfoRequest {
 
 export const postMyPageUserInfo = async (form: MyPageUserInfoRequest) => {
     try {
+        const cookieStore = await cookies();
+        const accessToken = cookieStore.get("accessToken")?.value;
+
+
         const res = await fetch(`${apiUrl}/api/v1/auth/my`, {
-            method: "POST",
+            method: "PUT",
             headers: {
                 "Content-Type": "application/json",
+                Authorization: `Bearer ${accessToken}`,
             },
             body: JSON.stringify(form),
             credentials: "include",
         });
 
         if (!res.ok) {
-            return {
-                error: "MyPage UserInfo Api 응답 에러",
-                status: res.status,
-                statusText: res.statusText,
-            };
+            throw new Error("MyPage UserInfo Api 응답 에러");
         }
 
         const result: MyPageUserInfo = await res.json();
